refactor viewer backend
Use a feature flag (`developer-mode`) to control initialization of
the livereload layer and whether to bake-in the static assets or
load dynamically from disk.

Moves echarts library into assets.

Splits plot related components into its own module.
