+++ Cargo.toml
@@ -64,6 +64,9 @@ nvml-wrapper = "0.10.0"
 [target.'cfg(target_os = "linux")'.build-dependencies]
 libbpf-cargo = { version = "0.24.8" }
 
+[features]
+developer-mode = []
+
 [profile.bench]
 debug = true
 lto = true
