+++ src/viewer/assets/index.html
@@ -4,7 +4,7 @@
     <title>Rezolus</title>
 
     <script src="lib/mithril.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/echarts/5.4.3/echarts.min.js"></script>
+    <script src="lib/charts/echarts.min.js"></script>
 
     <!-- Use type="module" to enable ES module imports -->
     <script type="module" src="lib/script.js"></script>
